refactor(core): extract disabled node filtering out of `findSubgraph`
I need a generic filter function for removing disabled nodes for running
the chat trigger.
