+++ packages/core/src/PartialExecutionUtils/index.ts
@@ -5,3 +5,4 @@ export { findSubgraph } from './findSubgraph';
 export { recreateNodeExecutionStack } from './recreateNodeExecutionStack';
 export { cleanRunData } from './cleanRunData';
 export { handleCycles } from './handleCycles';
+export { filterDisabledNodes } from './filterDisabledNodes';
